requirements.txt
fastapi
uvicorn
SQLAlchemy
asyncpg
psycopg2-binary
sqlacodegen==3.0.0rc5
python-dotenv==1.0.1